refactor: flatten core crate
